% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allometric_equations.R
\name{allometric_equations}
\alias{allometric_equations}
\title{Allometric Equations for Forest Simulation}
\description{
This file contains user-modifiable allometric equations for calculating
tree attributes from diameter at breast height (DBH). Users can customize
these equations for different species and forest types.
}
\keyword{internal}
