% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_simulation.R
\name{calc_canopy_cover}
\alias{calc_canopy_cover}
\title{Calculate total canopy cover accounting for overlap}
\usage{
calc_canopy_cover(x, y, crown_radius, plot_size = 100)
}
\arguments{
\item{x}{Vector of x coordinates (m)}

\item{y}{Vector of y coordinates (m)}

\item{crown_radius}{Vector of crown radii (m)}

\item{plot_size}{Size of plot (assumes square, m)}
}
\value{
Proportion of plot covered by canopy (0-1)
}
\description{
Uses a raster approach (0.5 m resolution) to account for crown overlap.
}
\examples{
# Small stand on a 20 m plot
x <- c(5, 10, 15)
y <- c(5, 10, 15)
cr <- c(2, 3, 2.5)
calc_canopy_cover(x, y, cr, plot_size = 20)
}
