% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_utils.R
\name{calc_canopy_cover_fast}
\alias{calc_canopy_cover_fast}
\title{Fast Canopy Cover Calculation (Vectorized)}
\usage{
calc_canopy_cover_fast(x, y, crown_radius, plot_size = 100, grid_res = 0.5)
}
\arguments{
\item{x}{Vector of x coordinates (m)}

\item{y}{Vector of y coordinates (m)}

\item{crown_radius}{Vector of crown radii (m)}

\item{plot_size}{Size of plot (m)}

\item{grid_res}{Grid resolution (m). Default 0.5m.}
}
\value{
Proportion of plot covered by canopy (0-1)
}
\description{
Optimized canopy cover calculation using matrix operations instead of loops.
Approximately 2-3x faster than original implementation.
}
\examples{
set.seed(1)
x <- runif(20, 0, 20)
y <- runif(20, 0, 20)
cr <- runif(20, 1, 3)
calc_canopy_cover_fast(x, y, cr, plot_size = 20)
}
