% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_utils.R
\name{calc_clark_evans_fast}
\alias{calc_clark_evans_fast}
\title{Fast Spatial Pattern Metrics}
\usage{
calc_clark_evans_fast(x, y, plot_size)
}
\arguments{
\item{x}{Vector of x coordinates}

\item{y}{Vector of y coordinates}

\item{plot_size}{Plot size (m)}
}
\value{
Clark-Evans R value
}
\description{
Optimized calculation of Clark-Evans R using efficient distance computation.
}
\keyword{internal}
