% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_simulation.R
\name{calc_tree_attributes}
\alias{calc_tree_attributes}
\title{Calculate all tree attributes from basic measurements}
\usage{
calc_tree_attributes(trees)
}
\arguments{
\item{trees}{Data table with x, y, Species, DBH columns}
}
\value{
Data table with all attributes
}
\description{
Calculate all tree attributes from basic measurements
}
