% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allometric_equations.R
\name{get_ponderosa_allometric_params}
\alias{get_ponderosa_allometric_params}
\title{Get default allometric parameters for ponderosa pine forest}
\usage{
get_ponderosa_allometric_params()
}
\value{
List of allometric parameters for ponderosa pine
}
\description{
Example parameter set for ponderosa pine dominated forests.
Demonstrates how to create custom allometric relationships.
}
\examples{
# Use ponderosa pine parameters
params <- get_ponderosa_allometric_params()
}
