% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_simulation.R
\name{perturb_add}
\alias{perturb_add}
\title{Add a new tree}
\usage{
perturb_add(trees, plot_size, species_names, species_probs, dbh_mean, dbh_sd)
}
\arguments{
\item{trees}{data.table. Tree data}

\item{plot_size}{Numeric. Plot dimension (m)}

\item{species_names}{Character vector. Available species}

\item{species_probs}{Numeric vector. Species proportions}

\item{dbh_mean}{Numeric. Mean DBH (cm)}

\item{dbh_sd}{Numeric. SD of DBH (cm)}
}
\value{
data.table. Modified tree data
}
\description{
Add a new tree
}
