% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_simulation.R
\name{perturb_move}
\alias{perturb_move}
\title{Move a random tree to a new location}
\usage{
perturb_move(trees, plot_size = 100)
}
\arguments{
\item{trees}{data.table. Tree data}

\item{plot_size}{Numeric. Plot dimension (m)}
}
\value{
data.table. Modified tree data
}
\description{
Move a random tree to a new location
}
\examples{
library(data.table)
trees <- data.table(Number = 1:5, x = 1:5, y = 1:5,
                    Species = "PIED", DBH = rep(20, 5))
set.seed(1)
perturb_move(trees, plot_size = 20)
}
