% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_simulation.R
\name{perturb_remove}
\alias{perturb_remove}
\title{Remove a random tree}
\usage{
perturb_remove(trees, min_trees = 10)
}
\arguments{
\item{trees}{data.table. Tree data}

\item{min_trees}{Integer. Minimum trees to maintain}
}
\value{
data.table. Modified tree data
}
\description{
Remove a random tree
}
