% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_simulation.R
\name{perturb_species}
\alias{perturb_species}
\title{Change species of a random tree}
\usage{
perturb_species(trees, species_names, species_probs)
}
\arguments{
\item{trees}{data.table. Tree data}

\item{species_names}{Character vector. Available species codes}

\item{species_probs}{Numeric vector. Target species proportions}
}
\value{
data.table. Modified tree data
}
\description{
Change species of a random tree
}
