% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_simulation.R
\name{plot_progress}
\alias{plot_progress}
\title{Plot progress during simulation - showing objective progress}
\usage{
plot_progress(
  trees,
  metrics,
  targets,
  history,
  iter,
  energy,
  temperature,
  plot_size,
  save_plot = FALSE
)
}
\arguments{
\item{trees}{Current trees data}

\item{metrics}{Current metrics}

\item{targets}{Target parameters}

\item{history}{History data table}

\item{iter}{Current iteration}

\item{energy}{Current energy}

\item{temperature}{Current temperature}

\item{plot_size}{Size of the plot area}

\item{save_plot}{Whether to save plot to file}
}
\description{
Plot progress during simulation - showing objective progress
}
