% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_system.R
\name{print_config}
\alias{print_config}
\title{Print Configuration Summary}
\usage{
print_config(config)
}
\arguments{
\item{config}{Configuration list from pj_huffman_2009() or create_config()}
}
\description{
Display a formatted summary of simulation configuration.
}
\examples{
config <- pj_huffman_2009()
print_config(config)
}
