% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_system.R
\name{save_config}
\alias{save_config}
\title{Export Configuration to File}
\usage{
save_config(config, file = "simulation_config.R")
}
\arguments{
\item{config}{Configuration list}

\item{file}{Output file path (default: "simulation_config.R")}
}
\description{
Save configuration as R script for reproducibility.
}
\examples{
config <- pj_huffman_2009(density_ha = 1000)
save_config(config, "my_simulation.R")
}
