% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_simulation.R
\name{calc_stand_metrics}
\alias{calc_stand_metrics}
\title{Calculate all stand-level metrics}
\usage{
calc_stand_metrics(trees, plot_size = 100)
}
\arguments{
\item{trees}{Data table with all tree attributes}

\item{plot_size}{Plot size (m)}
}
\value{
List of metrics
}
\description{
Calculate all stand-level metrics
}
