% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_simulation.R
\name{calc_tree_attributes}
\alias{calc_tree_attributes}
\title{Calculate all tree attributes from basic measurements}
\usage{
calc_tree_attributes(trees)
}
\arguments{
\item{trees}{Data table with x, y, Species, DBH columns}
}
\value{
Data table with all attributes added
}
\description{
Adds Height, CrownRadius, CrownDiameter, CrownArea, CrownBaseHeight,
CrownLength, and CanopyFuelMass columns to a tree data table.
}
\examples{
library(data.table)
trees <- data.table(
  Number = 1:3, x = c(5, 10, 15), y = c(5, 10, 15),
  Species = c("PIED", "JUMO", "PIED"), DBH = c(15, 20, 25)
)
result <- calc_tree_attributes(trees)
names(result)
}
