% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_simulation.R
\name{perturb_dbh}
\alias{perturb_dbh}
\title{Adjust DBH of a random tree}
\usage{
perturb_dbh(trees, dbh_sd_perturb = 3)
}
\arguments{
\item{trees}{data.table. Tree data}

\item{dbh_sd_perturb}{Numeric. SD of DBH perturbation (cm)}
}
\value{
data.table. Modified tree data
}
\description{
Adjust DBH of a random tree
}
\examples{
library(data.table)
trees <- data.table(Number = 1:5, x = 1:5, y = 1:5,
                    Species = "PIED", DBH = rep(20, 5))
set.seed(1)
perturb_dbh(trees, dbh_sd_perturb = 3)
}
