% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_simulation.R
\name{perturb_species}
\alias{perturb_species}
\title{Change species of a random tree}
\usage{
perturb_species(trees, species_names, species_probs)
}
\arguments{
\item{trees}{data.table. Tree data}

\item{species_names}{Character vector. Available species codes}

\item{species_probs}{Numeric vector. Target species proportions}
}
\value{
data.table. Modified tree data
}
\description{
Change species of a random tree
}
\examples{
library(data.table)
trees <- data.table(Number = 1:5, x = 1:5, y = 1:5,
                    Species = "PIED", DBH = rep(20, 5))
set.seed(1)
perturb_species(trees, c("PIED", "JUSO"), c(0.7, 0.3))
}
