% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_simulation.R
\name{plot_simulation_results}
\alias{plot_simulation_results}
\title{Plot Simulation Results}
\usage{
plot_simulation_results(result)
}
\arguments{
\item{result}{Result object from simulate_stand()}
}
\description{
Plot Simulation Results
}
